fix: prevent Let Encrypt flow break when CNAME/alternative fields are not provided
